+++ android/app/src/main/java/com/digginroom/digginroom/data/datasource/remote/AccountRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.data.source
+package com.digginroom.digginroom.data.datasource.remote
 
 import com.digginroom.digginroom.data.entity.JoinRequest
 import com.digginroom.digginroom.data.entity.JoinResponse
