+++ android/app/src/main/java/com/digginroom/digginroom/views/activity/RoomActivity.kt
@@ -1,21 +1,24 @@
-package com.digginroom.digginroom
+package com.digginroom.digginroom.views.activity
 
+import android.content.Context
+import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import com.digginroom.digginroom.data.datasource.RoomRemoteDataSource
+import com.digginroom.digginroom.R
+import com.digginroom.digginroom.data.datasource.remote.RoomRemoteDataSource
 import com.digginroom.digginroom.data.repository.DefaultRoomRepository
-import com.digginroom.digginroom.databinding.ActivityMainBinding
+import com.digginroom.digginroom.databinding.ActivityRoomBinding
 import com.digginroom.digginroom.viewmodels.RoomViewModel
 import com.digginroom.digginroom.viewmodels.YoutubeRoomCacheStrategy
 import com.digginroom.model.room.Room
 import com.digginroom.model.room.Song
 
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
+class RoomActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityRoomBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = DataBindingUtil.setContentView(this, R.layout.activity_main)
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_room)
         binding.lifecycleOwner = this
         binding.roomViewModel = RoomViewModel(
             mutableListOf(
@@ -50,5 +53,12 @@ class MainActivity : AppCompatActivity() {
                 RoomRemoteDataSource()
             )
         )
+    }
+
+    companion object {
+        fun start(context: Context) {
+            val intent = Intent(context, RoomActivity::class.java)
+            context.startActivity(intent)
+        }
     }
 }
