+++ android/app/src/main/java/com/digginroom/digginroom/views/bindingAdapter/RoomBindingAdapter.kt
@@ -1,7 +1,7 @@
-package com.digginroom.digginroom.views.bindingAdapters
+package com.digginroom.digginroom.views.bindingAdapter
 
 import androidx.databinding.BindingAdapter
-import com.digginroom.digginroom.views.RoomYoutubeView
+import com.digginroom.digginroom.views.customView.roomView.RoomYoutubeView
 import com.digginroom.model.room.Room
 
 object RoomBindingAdapter {
