+++ android/app/src/main/java/com/digginroom/digginroom/views/customView/roomView/RoomView.kt
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.views
+package com.digginroom.digginroom.views.customView.roomView
 
 import com.digginroom.model.room.Room
