+++ android/app/src/main/java/com/digginroom/digginroom/views/model/RoomModel.kt
@@ -1,0 +1,5 @@
+package com.digginroom.digginroom.views.model
+
+data class RoomModel(
+    val videoId: Int
+)
