--- android/app/src/main/java/com/digginroom/model/TempUiModel.kt
@@ -1,6 +1,0 @@
-package com.digginroom.model
-
-data class TempUiModel(
-    // 임시 룸 UI 데이터
-    val videoId: Int,
-)
