+++ android/app/src/main/res/layout/activity_join.xml
@@ -16,7 +16,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#110B2E"
-        tools:context=".views.JoinActivity">
+        tools:context=".views.activity.JoinActivity">
 
         <ProgressBar
             android:id="@+id/join_pb"
@@ -39,7 +39,7 @@
 
         <TextView
             android:id="@+id/join_tv_id_form_error"
-            isVisible="@{!viewModel.isValidId()}"
+            android:visibility="@{viewModel.isValidId ? View.INVISIBLE : View.VISIBLE}"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="14dp"
@@ -116,7 +116,7 @@
 
         <TextView
             android:id="@+id/join_tv_password_form_error"
-            isVisible="@{!viewModel.isValidPassword()}"
+            android:visibility="@{viewModel.isValidPassword ? View.INVISIBLE : View.VISIBLE}"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="14dp"
@@ -153,13 +153,12 @@
 
         <TextView
             android:id="@+id/join_tv_password_not_equal"
-            isVisible="@{!viewModel.isEqualPassword()}"
+            android:visibility="@{viewModel.isEqualPassword ? View.INVISIBLE : View.VISIBLE}"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="14dp"
             android:text="@string/join_password_not_equal_description"
             android:textColor="#AF1B1B"
-            android:visibility="invisible"
             app:layout_constraintStart_toStartOf="@+id/join_et_re_input_password"
             app:layout_constraintTop_toBottomOf="@+id/join_et_re_input_password" />
